ncklj

+++ SERVER_whisper_extraction.py
@@ -8,6 +8,7 @@ import time
 logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s - %(message)s')
 
 
+
 # Функция для преобразования аудио в текст
 def audio_to_text(audio_path, model):
     logging.info("Converting audio to text...")
@@ -16,8 +17,7 @@ def audio_to_text(audio_path, model):
     except Exception as e:
         logging.error(f"Whisper model error: {e}")
         return None
-    return result['text'] if result['text'].strip() != '' else None
-
+    return result['text'] if result['text'].strip() else None
 
 # Функция для извлечения аудио и транскрибации
 def encode_and_transcribe(video_path, model):
@@ -25,25 +25,27 @@ def encode_and_transcribe(video_path, model):
         logging.error("Video file does not exist")
         return None, 0
 
+    audio_path = "temp_audio.wav"
     start_time = time.time()
 
     try:
         # Извлечение аудиодорожки из видеофайла
         video = VideoFileClip(video_path)
-        audio_path = "temp_audio.wav"
         video.audio.write_audiofile(audio_path)
         video.close()
 
         # Транскрибация аудио
         transcription = audio_to_text(audio_path, model)
-
-        # Удаление временного аудиофайла
-        os.remove(audio_path)
-
     except Exception as e:
         logging.error(f"Error processing video: {e}")
         return None, 0
+    finally:
+        # Удаление временного аудиофайла, если он существует
+        if os.path.exists(audio_path):
+            os.remove(audio_path)
 
     processing_time = time.time() - start_time
     return transcription, processing_time
 
+# Пример вызова функции
+# encode_and_transcribe('path_to_video.mp4', whisper_model)
